Distribute SpannerFragment definitions and formalize enum
