+++ src/rendering/spanners.ts
@@ -1,7 +1,7 @@
-import { Beam, BeamRendering } from './beam';
-import { Slur, SlurRendering } from './slur';
-import { Tuplet, TupletRendering } from './tuplet';
-import * as vexflow from 'vexflow';
+import { Beam, BeamFragment, BeamRendering } from './beam';
+import { Slur, SlurFragment, SlurRendering } from './slur';
+import { Tuplet, TupletFragment, TupletRendering } from './tuplet';
+import { WedgeFragment, WedgeRendering } from './wedge';
 
 /** The result of rendering spanners. */
 export type SpannersRendering = {
@@ -9,10 +9,8 @@ export type SpannersRendering = {
   beams: BeamRendering[];
   tuplets: TupletRendering[];
   slurs: SlurRendering[];
+  wedges: WedgeRendering[];
 };
-
-/** Describes what part of the spanner lifecycle a spanner fragment is in. */
-export type SpannerFragmentPhase = 'unspecified' | 'start' | 'continue' | 'stop';
 
 /**
  * Represents a piece of a spanner.
@@ -24,45 +22,7 @@ export type SpannerFragmentPhase = 'unspecified' | 'start' | 'continue' | 'stop'
  *   - tuplets
  *   - slurs
  */
-export type SpannerFragment = BeamFragment | TupletFragment | SlurFragment;
-
-/** Represents a piece of a beam. */
-export type BeamFragment = {
-  type: 'beam';
-  phase: SpannerFragmentPhase;
-  vexflow: {
-    stemmableNote: vexflow.StemmableNote;
-  };
-};
-
-/** Represents a piece of a tuplet. */
-export type TupletFragment =
-  | {
-      type: 'tuplet';
-      phase: 'start';
-      vexflow: {
-        note: vexflow.Note;
-        location: vexflow.TupletLocation;
-      };
-    }
-  | {
-      type: 'tuplet';
-      phase: 'unspecified' | 'stop';
-      vexflow: {
-        note: vexflow.Note;
-      };
-    };
-
-/** Represents a piece of a slur. */
-export type SlurFragment = {
-  type: 'slur';
-  phase: SpannerFragmentPhase;
-  slurNumber: number;
-  vexflow: {
-    note: vexflow.Note;
-    keyIndex: number;
-  };
-};
+export type SpannerFragment = BeamFragment | TupletFragment | SlurFragment | WedgeFragment;
 
 /**
  * Houses **all** spanner structures that contain multiple notes.
@@ -85,6 +45,7 @@ export class Spanners {
       beams: this.getBeams().map((beam) => beam.render()),
       tuplets: this.getTuplets().map((tuplet) => tuplet.render()),
       slurs: this.getSlurs().map((slur) => slur.render()),
+      wedges: [],
     };
   }
